Add binary version of SILK lib, compiled for iOS with older version of XCode. Since XCode 5, apple dropped support for GCC, which means we have to compile with clang. Clang, sadly, doesn't support old-style ARM assembly, so we have to compile SILK in non-optimized fashion.
This is too much for armV7 arch, which chokes on silk_encode() and makes it impossible to use the codec for VoIP.
We'll be using this binary version of the lib, compiled with gcc under XCode 4.X, from now on. This makes the iOS calls possible again.

On the bad side, we now have unaligned functions and non-PIE code warnings in Xcode 5...